scc: exit on error
If there is an error in the processing of the input files, scc
should exit and inform the user.

scc is executed on a combined/preprocessed file and as a result
it doesn't have the granularity to see each input file individually.

Rather than moving preprocessing into scc (from spp), we can trap
the line number of the error and dump context around the line. This
gives the user a pointer to the input file and the specific line
that caused the problem.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>