patch: do not assume a branch specific patch queue is needed
When processing input files per-branch and global patch queues are
generated. If the meta-data has not created any branches in the repo,
no branch specific queue is required.

The tools assumed that one is always valid, and hence would throw a
non-zero exit code and stop processing.

By testing for a named per-branch queue, we avoid this issue.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>