scc: move unused patch queue under output dir
A placeholder "unused.patch.queue" is created for runs that do
not have patches, or runs that have patch generation disabled.
That output directory was being created in /tmp, which is fine,
except that other outputs use dirname of the current queue and
hence follow it to /tmp.

We fix this by putting it in $outdir like the rest of the
artifacts, and it is safely away from any shared directory
(and subsequent permissions issues).

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>