kern-tools: re-enable merge.cmd
Previous incarnations of the tools had a merge command that was
capable of merging feature branches into the build branch. This
support was removed to streamline processing of the tree.

Now that we have the .queue implementation, we can generate a
merge.queue from any merge commands found in .scc files and
allow the do_patch phase to take care of updating the tree.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>