kgit-s2q: add --clean to remove patching artifacts
Once patching has completed the last patch applied is stored in the file
kgit-s2q.last. If for whatever reason patching is called again, patching
resumes from that point.

Without manually removing that file there is currently no way to remove
the sentinel file if you want patching to resume from the start of the
series.

We add a --clean option to remove the file and keep the internal
structure of tracking hidden.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>