kgit-s2q: execute --clean and exit
The --clean flag to kgit-s2q exists to remove any fence post / sentinel
files that may be around from a previous run. That means --clean should
remove those files and then exit.

There were two issues with this, --clean was excuted too late in the
script AND it didn't exit after cleaning. This meant that if --clean
was issues before the tree was prepared, directories would be missing
and an error incorrectly thrown.

We move the clean operation to the top of the script and exit after
it is done and these issues are gone.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>