spp: only return files that match KMACHINE and KTYPE
The search utility of spp was incorrect and was returning files
that matched only a defined ktype.

This leads to the system potentially building the wrong BSP, and
not being able to report an error.

We fix the search to only return files that match both ktype and
kmachine, as well as return 0/1 for success/fail in the search.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>