kconf_check: allow non-hardware options to override hardware
If an option is specified as 'hardware' it means that it is critical
to board operation and as such, will be reported as a warning during
the audit phase if the option does not match the specified value.

Once something has been specified as hardware, it can't be undone,
which means that an understood mismatch cannot have the warning
inhibited.

This change allows an option that is later specified as non-hardware
to inhibit a warning for an option that was previous specified as
hardware.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>