spp: make fuzzy matching optional
Add a flag that can be used to toggle wether or not a partial
match is an error.

  --fuzz

When passed, partial patching will be used. If not passed the
default is to return nothing (which can be interpreted as an
error by the calling routines) if both the kernel type and
machine do not match.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>