tools/merge_config.sh: add CR after each fragment
If a fragment file doesn't contain a CR at the end, two config options
may be merged on the same line in the result file, leading to misconfiguration.

This patch adds a CR after each fragment to ensure that config options
are well separated in the result file.

Bug-AGL: SPEC-1475

Signed-off-by: Stephane Desneux <stephane.desneux@iot.bzh>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>