spp: fix end of line processing
When spp is including a file, read returns a failure result if the
last line has no ending. While this is posix behaviour, we do want that
line.

As such, we both check the read return code AND a variable for non
zero content to ensure that we have the entire file included.

Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>