symbol_why: last classification dictates option type
Being in both h/w and non-hardware leads us to results where we
can't easily silence a warning. So if we've added it to h/w we need
remove it from non-hardware. If we do it in both the hardware and
non hardware cases, the end result is that the last classification
wins, which is what we want.

Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>