kgit-config: BSP template and auto meta-data creation
With this version of kgit-config, we can run without an existing
meta-data repository and create a rough guideline of hardware
versus policy options using existing kgit-tooling. This isn't
as good as the yocto kernel-cache, but it is a starting point.

We can also now write BSP .scc files and hardware/policy fragments.

An example run follows:

 % kgit-config create --bsp mybsp --savetemps --ksrc \
    ~/poky-kernel/linux-yocto.git/ --defconfig \
    ~/poky-kernel/linux-yocto.git/arch/x86/configs/x86_64_defconfig --outdir .

Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>