kgit-s2q: allow author date to be used instead of current time/date
In situations where a buid needs to be reproducible, it is
sometimes desireable to use a patches author date, versus the
time when it is applied. This generates a consistent hash
between different patch applications.

We introduce a new option to enable this functionality, but
don't change the default (the application date).

Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>