spp: ensure that relocations and patch paths are absolute
If include directories (locations of layers) have relative
components, and the path to a patch doesn't share that relative
location, the detection of a common prefix can fail.

When the common prefix is incorrectly specified, it can be the
entire path to a patch. This results in everything being removed
and subtle errors (as this shouldn't happen).

By ensuring that both the relocation directories and patch path
are absolute, we can avoid this issue.

Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>