symbol_why: allow options to be flagged as built-in or module
Some options that are hardware related, can either be built in or
modules, and meet the criteria of being acceptable for a BSP.

To allow these options to not throw warnings, we create a new
configuration fragment to go along with the 'hardware.cfg' and
'nonhardware.cfg' special fragments.

This new fragment is called 'y_or_m_enabled.cfg' and is effective
for the fragments in its directory. The format is the same as
hardware/nonhardware.cfg, is simply a list of CONFIG_ options.

Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>