scc: only look for error in scc_output_file if it has valid content
When process_file function fails the output of the processed script is show to
the user, some parsing is performed as well to look for common errors so we
can point to the right input file.

This can only be done when the scc_output_file have some valid content
otherwise it will show invalid messages to the user.

Signed-off-by: Jose Quaresma <jose.quaresma@foundries.io>