Revert "spp: ensure that relocations and patch paths are absolute"
This reverts commit 9320c2a1aaa085e94abd15ede0d93ffdab578e9d.

The strip_common_prefix function is changed in a way that they will
use the expanded path of the input argument.
There are some changes in the process_file function for the *.cfg files
only and not none for *.scc, *.patch and *defconfig.

This doesn't work for cases where we have the defconfig files
provided in the some meta-layer and fail with:

[ERROR]: processing of file /tmp/tmp.8hcr45TMiO failed

The content of the '/tmp/tmp.8hcr45TMiO' file is currently:

| #
| # spp v0.8
| # processed: Thu Sep 29 04:15:36 PM UTC 2022
| #
| # This is a preprocessor output file, do not edit
| #
| #
| prefix
| kconf non-hardware /build/conf/../../layers/meta-lmp/meta-lmp-bsp/recipes-kernel/linux/linux-lmp-dev-mfgtool/imx8mq-evk/defconfig
| # run time: 0 seconds
| # processed files:
| # _cfg /build/conf/../../layers/meta-lmp/meta-lmp-bsp/recipes-kernel/linux/linux-lmp-dev-mfgtool/imx8mq-evk/defconfig

and with this patch applied (without 9320c2a):

| #
| # spp v0.8
| # processed: Thu Sep 29 04:18:21 PM UTC 2022
| #
| # This is a preprocessor output file, do not edit
| #
| #
| prefix /build/conf/../../layers/meta-lmp/meta-lmp-bsp/recipes-kernel/linux/linux-lmp-dev-mfgtool/imx8mq-evk/defconfig
| kconf non-hardware /build/conf/../../layers/meta-lmp/meta-lmp-bsp/recipes-kernel/linux/linux-lmp-dev-mfgtool/imx8mq-evk/defconfig
| # run time: 0 seconds
| # processed files:
| # _cfg /build/conf/../../layers/meta-lmp/meta-lmp-bsp/recipes-kernel/linux/linux-lmp-dev-mfgtool/imx8mq-evk/defconfig

so the 9320c2a patch removes the prefix in this use case.

Signed-off-by: Jose Quaresma <jose.quaresma@foundries.io>